+++ zed_components/src/zed_camera/src/zed_camera_component.cpp
@@ -17,6 +17,7 @@
 #include <vector>
 #include <limits>
 #include <sys/resource.h>
+#include <stdexcept>
 
 #include "zed_camera_component.hpp"
 #include "sl_logging.hpp"
@@ -71,6 +72,7 @@ ZedCamera::ZedCamera(const rclcpp::NodeOptions & options)
   mBodyTrkQos(1),
   mClickedPtQos(1),
   mGnssFixQos(1),
+  mClockQos(1),
   mAiInstanceID(0),
   mDiagUpdater(this),
   mImuTfFreqTimer(get_clock()),
@@ -95,7 +97,8 @@ ZedCamera::ZedCamera(const rclcpp::NodeOptions & options)
   mLastTs_odom(TIMEZERO_ROS),
   mLastTs_pose(TIMEZERO_ROS),
   mLastTs_pc(TIMEZERO_ROS),
-  mPrevTs_pc(TIMEZERO_ROS)
+  mPrevTs_pc(TIMEZERO_ROS),
+  mLastClock(TIMEZERO_ROS)
 {
   RCLCPP_INFO(get_logger(), "********************************");
   RCLCPP_INFO(get_logger(), "      ZED Camera Component ");
@@ -151,6 +154,7 @@ ZedCamera::~ZedCamera()
   // ----> Stop subscribers
   mClickedPtSub.reset();
   mGnssFixSub.reset();
+  mClockSub.reset();
   // <---- Stop subscribers
 
   if (mObjDetRunning) {
@@ -493,6 +497,9 @@ void ZedCamera::getDebugParams()
   getParam("debug.debug_common", mDebugCommon, mDebugCommon);
   RCLCPP_INFO(get_logger(), " * Debug Common: %s", mDebugCommon ? "TRUE" : "FALSE");
 
+  getParam("debug.debug_sim", mDebugSim, mDebugSim);
+  RCLCPP_INFO(get_logger(), " * Debug Simulation: %s", mDebugSim ? "TRUE" : "FALSE");
+
   getParam("debug.debug_video_depth", mDebugVideoDepth, mDebugVideoDepth);
   RCLCPP_INFO(get_logger(), " * Debug Video/Depth: %s", mDebugVideoDepth ? "TRUE" : "FALSE");
 
@@ -531,7 +538,7 @@ void ZedCamera::getDebugParams()
   getParam("debug.debug_advanced", mDebugAdvanced, mDebugAdvanced);
   RCLCPP_INFO(get_logger(), " * Debug Advanced: %s", mDebugAdvanced ? "TRUE" : "FALSE");
 
-  mDebugMode = mDebugCommon || mDebugVideoDepth || mDebugCamCtrl || mDebugPointCloud ||
+  mDebugMode = mDebugCommon || mDebugSim || mDebugVideoDepth || mDebugCamCtrl || mDebugPointCloud ||
     mDebugPosTracking || mDebugGnss || mDebugSensors || mDebugMapping ||
     mDebugObjectDet || mDebugBodyTrk || mDebugAdvanced;
 
@@ -559,15 +566,24 @@ void ZedCamera::getDebugParams()
 void ZedCamera::getSimParams()
 {
   // SIMULATION active?
-  if (!get_parameter("use_sim_time", mSimEnabled)) {
-    RCLCPP_WARN_STREAM(
+  getParam("simulation.sim_enabled", mSimMode, mSimMode);
+
+  if (!get_parameter("use_sim_time", mUseSimTime)) {
+    RCLCPP_WARN(
       get_logger(),
-      "The parameter 'use_sim_time' is not available or is not valid, using the default value: " <<
-        mSimEnabled);
+      "The parameter 'use_sim_time' is not available or is not valid, using the default value.");
   }
-  if (mSimEnabled) {
+
+  if (mSimMode) {
     RCLCPP_INFO(get_logger(), " *** SIMULATION MODE ACTIVE ***");
-    getParam("sim_address", mSimAddr, mSimAddr, " * Simulator address: ");
+    getParam("simulation.sim_address", mSimAddr, mSimAddr, " * Sim. server address: ");
+    getParam("simulation.sim_port", mSimPort, mSimPort, " * Sim. server port: ");
+
+    RCLCPP_INFO_STREAM(get_logger(), " * Use Sim Time: " << (mUseSimTime ? "TRUE" : "FALSE"));
+  } else if (mUseSimTime) {
+    RCLCPP_WARN(
+      get_logger(),
+      "The parameter 'use_sim_time' is set to 'true', but simulation mode is not enabled. UNPREDICTABLE BEHAVIORS EXPECTED.");
   }
 }
 
@@ -610,7 +626,7 @@ void ZedCamera::getGeneralParams()
       RCLCPP_INFO_STREAM(
         get_logger(), " + Playing an SVO for " << sl::toString(
           mCamUserModel) << " camera model.");
-    } else if (mSimEnabled) {
+    } else if (mSimMode) {
       RCLCPP_INFO_STREAM(
         get_logger(), " + Simulating a " << sl::toString(
           mCamUserModel) << " camera model.");
@@ -626,7 +642,7 @@ void ZedCamera::getGeneralParams()
       RCLCPP_INFO_STREAM(
         get_logger(), " + Playing an SVO for " << sl::toString(
           mCamUserModel) << " camera model.");
-    } else if (mSimEnabled) {
+    } else if (mSimMode) {
       RCLCPP_INFO_STREAM(
         get_logger(), " + Simulating a " << sl::toString(
           mCamUserModel) << " camera model.");
@@ -644,55 +660,65 @@ void ZedCamera::getGeneralParams()
 
   getParam("general.sdk_verbose", mVerbose, mVerbose, " * SDK Verbose: ");
   getParam("general.camera_name", mCameraName, mCameraName, " * Camera name: ");
-  getParam("general.zed_id", mCamId, mCamId, " * Camera ID: ");
   getParam("general.serial_number", mCamSerialNumber, mCamSerialNumber, " * Camera SN: ");
   getParam(
     "general.camera_timeout_sec", mCamTimeoutSec, mCamTimeoutSec, " * Camera timeout [sec]: ");
   getParam(
     "general.camera_max_reconnect", mMaxReconnectTemp, mMaxReconnectTemp,
     " * Camera reconnection temptatives: ");
-  getParam(
-    "general.grab_frame_rate", mCamGrabFrameRate, mCamGrabFrameRate, " * Camera framerate: ");
+  if (mSimMode) {
+    RCLCPP_INFO(get_logger(), "* [Simulation mode] Camera framerate forced to 60 Hz");
+    mCamGrabFrameRate = 60;
+  } else {
+    getParam(
+      "general.grab_frame_rate", mCamGrabFrameRate, mCamGrabFrameRate,
+      " * Camera framerate: ");
+  }
   getParam("general.gpu_id", mGpuId, mGpuId, " * GPU ID: ");
 
   // TODO(walter) ADD SVO SAVE COMPRESSION PARAMETERS
 
-  std::string resol = "AUTO";
-  getParam("general.grab_resolution", resol, resol);
-  if (resol == "AUTO") {
-    mCamResol = sl::RESOLUTION::AUTO;
-  } else if (sl_tools::isZEDX(mCamUserModel)) {
-    if (resol == "HD1200") {
-      mCamResol = sl::RESOLUTION::HD1200;
-    } else if (resol == "HD1080") {
-      mCamResol = sl::RESOLUTION::HD1080;
-    } else if (resol == "SVGA") {
-      mCamResol = sl::RESOLUTION::SVGA;
-    } else {
-      RCLCPP_WARN(
-        get_logger(),
-        "Not valid 'general.grab_resolution' value: '%s'. Using 'AUTO' setting.", resol.c_str());
-      mCamResol = sl::RESOLUTION::AUTO;
-    }
-    RCLCPP_INFO_STREAM(
-      get_logger(), " * Camera resolution: " << sl::toString(mCamResol).c_str());
+  if (mSimMode) {
+    RCLCPP_INFO(get_logger(), "* [Simulation mode] Camera resolution forced to 'HD1080'");
+    mCamResol = sl::RESOLUTION::HD1080;
   } else {
-    if (resol == "HD2K") {
-      mCamResol = sl::RESOLUTION::HD2K;
-    } else if (resol == "HD1080") {
-      mCamResol = sl::RESOLUTION::HD1080;
-    } else if (resol == "HD720") {
-      mCamResol = sl::RESOLUTION::HD720;
-    } else if (resol == "VGA") {
-      mCamResol = sl::RESOLUTION::VGA;
-    } else {
-      RCLCPP_WARN(
-        get_logger(),
-        "Not valid 'general.grab_resolution' value: '%s'. Using 'AUTO' setting.", resol.c_str());
+    std::string resol = "AUTO";
+    getParam("general.grab_resolution", resol, resol);
+    if (resol == "AUTO") {
       mCamResol = sl::RESOLUTION::AUTO;
+    } else if (sl_tools::isZEDX(mCamUserModel)) {
+      if (resol == "HD1200") {
+        mCamResol = sl::RESOLUTION::HD1200;
+      } else if (resol == "HD1080") {
+        mCamResol = sl::RESOLUTION::HD1080;
+      } else if (resol == "SVGA") {
+        mCamResol = sl::RESOLUTION::SVGA;
+      } else {
+        RCLCPP_WARN(
+          get_logger(),
+          "Not valid 'general.grab_resolution' value: '%s'. Using 'AUTO' setting.", resol.c_str());
+        mCamResol = sl::RESOLUTION::AUTO;
+      }
+      RCLCPP_INFO_STREAM(
+        get_logger(), " * Camera resolution: " << sl::toString(mCamResol).c_str());
+    } else {
+      if (resol == "HD2K") {
+        mCamResol = sl::RESOLUTION::HD2K;
+      } else if (resol == "HD1080") {
+        mCamResol = sl::RESOLUTION::HD1080;
+      } else if (resol == "HD720") {
+        mCamResol = sl::RESOLUTION::HD720;
+      } else if (resol == "VGA") {
+        mCamResol = sl::RESOLUTION::VGA;
+      } else {
+        RCLCPP_WARN(
+          get_logger(),
+          "Not valid 'general.grab_resolution' value: '%s'. Using 'AUTO' setting.", resol.c_str());
+        mCamResol = sl::RESOLUTION::AUTO;
+      }
+      RCLCPP_INFO_STREAM(
+        get_logger(), " * Camera resolution: " << sl::toString(mCamResol).c_str());
     }
-    RCLCPP_INFO_STREAM(
-      get_logger(), " * Camera resolution: " << sl::toString(mCamResol).c_str());
   }
 
   std::string out_resol = "MEDIUM";
@@ -729,16 +755,22 @@ void ZedCamera::getGeneralParams()
 
   // Dynamic parameters
 
-  getParam("general.pub_frame_rate", mPubFrameRate, mPubFrameRate, "", false);
-  if (mPubFrameRate > mCamGrabFrameRate) {
-    RCLCPP_WARN(get_logger(), "'pub_frame_rate' cannot be bigger than 'grab_frame_rate'");
-    mPubFrameRate = mCamGrabFrameRate;
+  if (mSimMode) {
+    RCLCPP_INFO(get_logger(), "* [Simulation mode] Publish framerate forced to 60 Hz");
+    mPubFrameRate = 60;
+  } else {
+    getParam("general.pub_frame_rate", mPubFrameRate, mPubFrameRate, "", false);
+    if (mPubFrameRate > mCamGrabFrameRate) {
+      RCLCPP_WARN(get_logger(), "'pub_frame_rate' cannot be bigger than 'grab_frame_rate'");
+      mPubFrameRate = mCamGrabFrameRate;
+    }
+    if (mPubFrameRate < 0.1) {
+      RCLCPP_WARN(get_logger(), "'pub_frame_rate' cannot be lower than 0.1 Hz or negative.");
+      mPubFrameRate = mCamGrabFrameRate;
+    }
+    RCLCPP_INFO_STREAM(get_logger(), " * [DYN] Publish framerate [Hz]:  " << mPubFrameRate);
   }
-  if (mPubFrameRate < 0.1) {
-    RCLCPP_WARN(get_logger(), "'pub_frame_rate' cannot be lower than 0.1 Hz or negative.");
-    mPubFrameRate = mCamGrabFrameRate;
-  }
-  RCLCPP_INFO_STREAM(get_logger(), " * [DYN] Publish framerate [Hz]:  " << mPubFrameRate);
+
 }
 
 void ZedCamera::getVideoParams()
@@ -3373,7 +3405,9 @@ void ZedCamera::initSubscribers()
       mClickedPtTopic, mClickedPtQos, std::bind(&ZedCamera::callback_clickedPoint, this, _1));
 
 
-    RCLCPP_INFO_STREAM(get_logger(), " * Plane detection: '" << mClickedPtTopic.c_str() << "'");
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Plane detection: '" << mClickedPtSub->get_topic_name() << "'");
   }
   // <---- Clicked Point Subscriber
 
@@ -3397,9 +3431,33 @@ void ZedCamera::initSubscribers()
     mGnssFixSub = create_subscription<sensor_msgs::msg::NavSatFix>(
       mGnssTopic, mGnssFixQos, std::bind(&ZedCamera::callback_gnssFix, this, _1));
 
-    RCLCPP_INFO_STREAM(get_logger(), " * GNSS Fix: '" << mGnssTopic.c_str() << "'");
+    RCLCPP_INFO_STREAM(get_logger(), " * GNSS Fix: '" << mGnssFixSub->get_topic_name() << "'");
   }
   // <---- GNSS Fix Subscriber
+
+  // ----> Clock Subscriber
+  /* From `$ ros2 topic info /clock -v`
+
+    QoS profile:
+      Reliability: RELIABLE
+      History (Depth): KEEP_LAST (10)
+      Durability: VOLATILE
+      Lifespan: Infinite
+      Deadline: Infinite
+      Liveliness: AUTOMATIC
+      Liveliness lease duration: Infinite
+  */
+
+  if (mUseSimTime) {
+    mClockAvailable = false;
+
+    mClockQos.reliability(RMW_QOS_POLICY_RELIABILITY_RELIABLE);
+    mClockQos.durability(RMW_QOS_POLICY_DURABILITY_VOLATILE);
+    mClockSub = create_subscription<rosgraph_msgs::msg::Clock>(
+      "/clock", mClockQos, std::bind(&ZedCamera::callback_clock, this, _1));
+
+    RCLCPP_INFO_STREAM(get_logger(), " * Sim Clock: '" << mClockSub->get_topic_name() << "'");
+  }
 }
 
 bool ZedCamera::startCamera()
@@ -3420,10 +3478,12 @@ bool ZedCamera::startCamera()
   // <---- TF2 Transform
 
   // ----> ZED configuration
-  if (mSimEnabled) {  // Simulation?
-    RCLCPP_INFO(get_logger(), "*** CONNECTING TO THE SIMULATOR ***");
+  if (mSimMode) {  // Simulation?
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      "*** CONNECTING TO THE SIMULATION SERVER [" << mSimAddr.c_str() << ":" << mSimPort <<
+        "] ***");
 
-    int mSimPort = ZED_SDK_PORT + 2 * mCamId;
     mInitParams.input.setFromStream(mSimAddr.c_str(), mSimPort);
   } else if (!mSvoFilepath.empty()) {
     RCLCPP_INFO(get_logger(), "*** SVO OPENING ***");
@@ -3437,9 +3497,7 @@ bool ZedCamera::startCamera()
     mInitParams.grab_compute_capping_fps = static_cast<float>(mPubFrameRate);
     mInitParams.camera_resolution = static_cast<sl::RESOLUTION>(mCamResol);
 
-    if (mCamSerialNumber == 0) {
-      mInitParams.input.setFromCameraID(mCamId);
-    } else {
+    if (mCamSerialNumber > 0) {
       mInitParams.input.setFromSerialNumber(mCamSerialNumber);
     }
   }
@@ -3466,10 +3524,8 @@ bool ZedCamera::startCamera()
 
   mThreadStop = false;
 
-  if (!mSvoMode && !mSimEnabled) {
-    if (mCamSerialNumber <= 0) {
-      mInitParams.input.setFromCameraID(mCamId);
-    } else {
+  if (!mSvoMode && !mSimMode) {
+    if (mCamSerialNumber > 0) {
       mInitParams.input.setFromSerialNumber(mCamSerialNumber);
     }
   }
@@ -3487,9 +3543,9 @@ bool ZedCamera::startCamera()
     if (mSvoMode) {
       RCLCPP_WARN(get_logger(), "Error opening SVO: %s", sl::toString(mConnStatus).c_str());
       return false;
-    } else if (mSimEnabled) {
+    } else if (mSimMode) {
       RCLCPP_WARN(
-        get_logger(), "Error connecting to the simulator: %s", sl::toString(
+        get_logger(), "Error connecting to the simulation server: %s", sl::toString(
           mConnStatus).c_str());
     } else {
       RCLCPP_WARN(get_logger(), "Error opening camera: %s", sl::toString(mConnStatus).c_str());
@@ -3587,7 +3643,7 @@ bool ZedCamera::startCamera()
   RCLCPP_INFO_STREAM(
     get_logger(),
     " * Focal Lenght -> " << camInfo.camera_configuration.calibration_parameters.left_cam.focal_length_metric <<
-      " m");
+      " mm");
 #endif
 
   RCLCPP_INFO_STREAM(
@@ -3907,7 +3963,7 @@ bool ZedCamera::startCamera()
 
   // Disable AEC_AGC and Auto Whitebalance to trigger it if user set it to
   // automatic
-  if (!mSvoMode && !mSimEnabled) {
+  if (!mSvoMode && !mSimMode) {
     mZed.setCameraSettings(sl::VIDEO_SETTINGS::AEC_AGC, 0);
     mZed.setCameraSettings(sl::VIDEO_SETTINGS::WHITEBALANCE_AUTO, 0);
     // Force parameters with a dummy grab
@@ -3918,8 +3974,12 @@ bool ZedCamera::startCamera()
   // ----> Timestamp
   if (mSvoMode) {
     mFrameTimestamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::CURRENT));
-  } else if (mSimEnabled) {
-    mFrameTimestamp = get_clock()->now(); // We must use the simulation time, not the SDK time
+  } else if (mSimMode) {
+    if (mUseSimTime) {
+      mFrameTimestamp = get_clock()->now();
+    } else {
+      mFrameTimestamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::IMAGE));
+    }
   } else {
     mFrameTimestamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::IMAGE));
   }
@@ -3970,7 +4030,7 @@ bool ZedCamera::startCamera()
     mFusionConfig =
       std::make_shared<sl::FusionConfiguration>();
 
-    if (mSimEnabled) {
+    if (mSimMode) {
       //TODO(Walter) Modify when support for streaming input is added in the SDK
       //mFusionConfig->input_type.setFromStream(mSimAddr, mSimPort);
       mFusionConfig->input_type.setFromSerialNumber(mCamSerialNumber);
@@ -4017,13 +4077,13 @@ bool ZedCamera::startCamera()
 void ZedCamera::initThreads()
 {
   // ----> Start CMOS Temperatures thread
-  if (!mSimEnabled && !sl_tools::isZED(mCamRealModel) && !sl_tools::isZEDM(mCamRealModel)) {
+  if (!mSimMode && !sl_tools::isZED(mCamRealModel) && !sl_tools::isZEDM(mCamRealModel)) {
     startTempPubTimer();
   }
   // <---- Start CMOS Temperatures thread
 
   // ----> Start Sensors thread if not sync
-  if (!mSvoMode && !mSensCameraSync && !sl_tools::isZED(mCamRealModel)) {
+  if (!mSimMode && !mSvoMode && !mSensCameraSync && !sl_tools::isZED(mCamRealModel)) {
     mSensThread = std::thread(&ZedCamera::threadFunc_pubSensorsData, this);
   }
   // <---- Start Sensors thread if not sync
@@ -4639,12 +4699,12 @@ bool ZedCamera::getCamera2BaseTransform()
   } catch (tf2::TransformException & ex) {
     if (!mCamera2BaseFirstErr) {
       rclcpp::Clock steady_clock(RCL_STEADY_TIME);
-      RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1.0, "Transform error: %s", ex.what());
+      RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1000.0, "Transform error: %s", ex.what());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0, "The tf from '%s' to '%s' is not available.",
+        get_logger(), steady_clock, 1000.0, "The tf from '%s' to '%s' is not available.",
         mCameraFrameId.c_str(), mBaseFrameId.c_str());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0,
+        get_logger(), steady_clock, 1000.0,
         "Note: one of the possible cause of the problem is the absense of an "
         "instance "
         "of the `robot_state_publisher` node publishing the correct static "
@@ -4703,12 +4763,12 @@ bool ZedCamera::getSens2CameraTransform()
   } catch (tf2::TransformException & ex) {
     if (!mSensor2CameraTransfFirstErr) {
       rclcpp::Clock steady_clock(RCL_STEADY_TIME);
-      RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1.0, "Transform error: %s", ex.what());
+      RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1000.0, "Transform error: %s", ex.what());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0, "The tf from '%s' to '%s' is not available.",
+        get_logger(), steady_clock, 1000.0, "The tf from '%s' to '%s' is not available.",
         mDepthFrameId.c_str(), mCameraFrameId.c_str());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0,
+        get_logger(), steady_clock, 1000.0,
         "Note: one of the possible cause of the problem is the absense of an "
         "instance "
         "of the `robot_state_publisher` node publishing the correct static "
@@ -4768,12 +4828,12 @@ bool ZedCamera::getSens2BaseTransform()
   } catch (tf2::TransformException & ex) {
     if (!mSensor2BaseTransfFirstErr) {
       rclcpp::Clock steady_clock(RCL_STEADY_TIME);
-      RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1.0, "Transform error: %s", ex.what());
+      RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1000.0, "Transform error: %s", ex.what());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0, "The tf from '%s' to '%s' is not available.",
+        get_logger(), steady_clock, 1000.0, "The tf from '%s' to '%s' is not available.",
         mDepthFrameId.c_str(), mBaseFrameId.c_str());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0,
+        get_logger(), steady_clock, 1000.0,
         "Note: one of the possible cause of the problem is the absense of an "
         "instance "
         "of the `robot_state_publisher` node publishing the correct static "
@@ -4832,12 +4892,12 @@ bool ZedCamera::getGnss2BaseTransform()
   } catch (tf2::TransformException & ex) {
     if (!mGnss2BaseTransfFirstErr) {
       rclcpp::Clock steady_clock(RCL_STEADY_TIME);
-      DEBUG_STREAM_THROTTLE_GNSS(1.0, "Transform error: " << ex.what());
+      DEBUG_STREAM_THROTTLE_GNSS(1000.0, "Transform error: " << ex.what());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0, "The tf from '%s' to '%s' is not available.",
+        get_logger(), steady_clock, 1000.0, "The tf from '%s' to '%s' is not available.",
         mGnssFrameId.c_str(), mBaseFrameId.c_str());
       RCLCPP_WARN_THROTTLE(
-        get_logger(), steady_clock, 1.0,
+        get_logger(), steady_clock, 1000.0,
         "Note: one of the possible cause of the problem is the absense of an "
         "instance "
         "of the `robot_state_publisher` node publishing the correct static "
@@ -5042,6 +5102,15 @@ void ZedCamera::threadFunc_zedGrab()
 
   // Infinite grab thread
   while (1) {
+
+    if (mUseSimTime && !mClockAvailable) {
+      rclcpp::Clock steady_clock(RCL_STEADY_TIME);
+      RCLCPP_WARN_THROTTLE(
+        get_logger(), steady_clock, 5000.0,
+        "Waiting for a valid simulation time on the '/clock' topic...");
+      continue;
+    }
+
     sl_tools::StopWatch grabElabTimer(get_clock());
 
     // ----> Interruption check
@@ -5064,7 +5133,7 @@ void ZedCamera::threadFunc_zedGrab()
     // <---- Apply depth settings
 
     // ----> Apply video dynamic parameters
-    if (!mSimEnabled && !mSvoMode) {
+    if (!mSimMode && !mSvoMode) {
       applyVideoSettings();
     }
     // <---- Apply video dynamic parameters
@@ -5187,14 +5256,18 @@ void ZedCamera::threadFunc_zedGrab()
       // ----> Timestamp
       if (mSvoMode) {
         mFrameTimestamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::CURRENT));
-      } else if (mSimEnabled) {
-        mFrameTimestamp = get_clock()->now(); // We must use the simulation time, not the SDK time
+      } else if (mSimMode) {
+        if (mUseSimTime) {
+          mFrameTimestamp = get_clock()->now();
+        } else {
+          mFrameTimestamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::IMAGE));
+        }
       } else {
         mFrameTimestamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::IMAGE));
       }
       // <---- Timestamp
 
-      if (!mSimEnabled) {
+      if (!mSimMode) {
         if (mGnssFusionEnabled && mGnssFixNew) {
           mGnssFixNew = false;
 
@@ -5225,7 +5298,7 @@ void ZedCamera::threadFunc_zedGrab()
 
         if (!mRecStatus.status) {
           rclcpp::Clock steady_clock(RCL_STEADY_TIME);
-          RCLCPP_ERROR_THROTTLE(get_logger(), steady_clock, 1.0, "Error saving frame to SVO");
+          RCLCPP_ERROR_THROTTLE(get_logger(), steady_clock, 1000.0, "Error saving frame to SVO");
         }
       }
       mRecMutex.unlock();
@@ -5242,7 +5315,7 @@ void ZedCamera::threadFunc_zedGrab()
       publishVideoDepth(pub_ts);
 
       if (!sl_tools::isZED(mCamRealModel) && mVdPublishing && pub_ts != TIMEZERO_ROS) {
-        if (mSensCameraSync || mSvoMode) {
+        if (mSensCameraSync || mSvoMode || mSimMode) {
           publishSensorsData(pub_ts);
         }
       }
@@ -5329,7 +5402,7 @@ void ZedCamera::threadFunc_zedGrab()
 rclcpp::Time ZedCamera::publishSensorsData(rclcpp::Time t)
 {
   // ----> Subscribers count
-  DEBUG_STREAM_ONCE_SENS("Sensors callback: counting subscribers");
+  DEBUG_STREAM_SENS("Sensors callback: counting subscribers");
 
   size_t imu_SubNumber = 0;
   size_t imu_RawSubNumber = 0;
@@ -5367,7 +5440,7 @@ rclcpp::Time ZedCamera::publishSensorsData(rclcpp::Time t)
 
   sl::SensorsData sens_data;
 
-  if (mSvoMode || mSensCameraSync || mSimEnabled) {
+  if (mSvoMode || mSensCameraSync || mSimMode) {
     sl::ERROR_CODE err = mZed.getSensorsData(sens_data, sl::TIME_REFERENCE::IMAGE);
     if (err != sl::ERROR_CODE::SUCCESS) {
       RCLCPP_WARN_STREAM(get_logger(), "sl::getSensorsData error: " << sl::toString(err).c_str());
@@ -5385,8 +5458,12 @@ rclcpp::Time ZedCamera::publishSensorsData(rclcpp::Time t)
     ts_imu = t;
     ts_baro = t;
     ts_mag = t;
-  } else if (mSimEnabled) {
-    ts_imu = get_clock()->now();
+  } else if (mSimMode) {
+    if (mUseSimTime) {
+      ts_imu = get_clock()->now();
+    } else {
+      ts_imu = sl_tools::slTime2Ros(sens_data.imu.timestamp);
+    }
     ts_baro = ts_imu;
     ts_mag = ts_imu;
   } else {
@@ -5411,7 +5488,7 @@ rclcpp::Time ZedCamera::publishSensorsData(rclcpp::Time t)
     return TIMEZERO_ROS;
   }
 
-  if (mSimEnabled) {
+  if (mSimMode) {
     new_baro_data = false;
     new_mag_data = false;
   }
@@ -6156,6 +6233,7 @@ void ZedCamera::retrieveVideoDepth()
 
 void ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
 {
+  DEBUG_VD("*** Publish Video and Depth topics *** ");
   sl_tools::StopWatch vdElabTimer(get_clock());
 
   // ----> Check RGB/Depth sync
@@ -6202,8 +6280,14 @@ void ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
   // <---- Check if a grab has been done before publishing the same images
 
   rclcpp::Time timeStamp;
-  if (mSvoMode || mSimEnabled) {
+  if (mSvoMode) {
     timeStamp = mFrameTimestamp;
+  } else if (mSimMode) {
+    if (mUseSimTime) {
+      timeStamp = get_clock()->now();
+    } else {
+      timeStamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::IMAGE));
+    }
   } else {
     timeStamp = sl_tools::slTime2Ros(mSdkGrabTS, get_clock()->get_clock_type());
   }
@@ -6357,6 +6441,8 @@ void ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
 
   mVdPubFreqTimer.tic();
   // <---- Check publishing frequency
+
+  DEBUG_VD("*** Video and Depth topics published *** ");
 }
 
 void ZedCamera::publishImageWithInfo(
@@ -6469,7 +6555,7 @@ void ZedCamera::processOdometry()
     }
   } else if (mFloorAlignment) {
     DEBUG_STREAM_THROTTLE_PT(
-      5.0,
+      5000.0,
       "Odometry will be published as soon as the floor as "
       "been detected for the first time");
   }
@@ -7813,9 +7899,15 @@ void ZedCamera::publishPointCloud()
 
   int ptsCount = width * height;
 
-  if (mSvoMode || mSimEnabled) {
+  if (mSvoMode) {
     // pcMsg->header.stamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::CURRENT));
     pcMsg->header.stamp = mFrameTimestamp;
+  } else if (mSimMode) {
+    if (mUseSimTime) {
+      pcMsg->header.stamp = mFrameTimestamp;
+    } else {
+      pcMsg->header.stamp = sl_tools::slTime2Ros(mMatCloud.timestamp);
+    }
   } else {
     pcMsg->header.stamp = sl_tools::slTime2Ros(mMatCloud.timestamp);
   }
@@ -8034,8 +8126,14 @@ void ZedCamera::callback_pubFusedPc()
         float * cloud_pts = reinterpret_cast<float *>(mFusedPC.chunks[c].vertices.data());
         memcpy(ptCloudPtr, cloud_pts, 4 * chunkSize * sizeof(float));
         ptCloudPtr += 4 * chunkSize;
-        if (mSvoMode || mSimEnabled) {
+        if (mSvoMode) {
           pointcloudFusedMsg->header.stamp = mFrameTimestamp;
+        } else if (mSimMode) {
+          if (mUseSimTime) {
+            pointcloudFusedMsg->header.stamp = mFrameTimestamp;
+          } else {
+            pointcloudFusedMsg->header.stamp = sl_tools::slTime2Ros(mFusedPC.chunks[c].timestamp);
+          }
         } else {
           pointcloudFusedMsg->header.stamp = sl_tools::slTime2Ros(mFusedPC.chunks[c].timestamp);
         }
@@ -8542,6 +8640,9 @@ void ZedCamera::callback_pauseSvoInput(
 
 void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWrapper & stat)
 {
+
+  DEBUG_COMM("*** Update Diagnostic ***");
+
   if (mConnStatus != sl::ERROR_CODE::SUCCESS) {
     stat.summary(diagnostic_msgs::msg::DiagnosticStatus::ERROR, sl::toString(mConnStatus).c_str());
     return;
@@ -8573,7 +8674,7 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
       stat.summary(diagnostic_msgs::msg::DiagnosticStatus::OK, "Camera grabbing");
     }
 
-    if (mSimEnabled) {
+    if (mSimMode) {
       stat.add("Input mode", "SIMULATION");
     } else if (mSvoMode) {
       stat.add("Input mode", "SVO");
@@ -8725,7 +8826,7 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
     stat.add("IMU Sensor", "Topics not subscribed");
   }
 
-  if (sl_tools::isZED2OrZED2i(mCamRealModel)) {
+  if (!mSvoMode && !mSimMode && sl_tools::isZED2OrZED2i(mCamRealModel)) {
     if (mMagPublishing) {
       double freq = 1. / mMagPeriodMean_sec->getAvg();
       stat.addf("Magnetometer", "Mean Frequency: %.1f Hz", freq);
@@ -8736,7 +8837,7 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
     stat.add("Magnetometer Sensor", "N/A");
   }
 
-  if (sl_tools::isZED2OrZED2i(mCamRealModel)) {
+  if (!mSvoMode && !mSimMode && sl_tools::isZED2OrZED2i(mCamRealModel)) {
     if (mBaroPublishing) {
       double freq = 1. / mBaroPeriodMean_sec->getAvg();
       stat.addf("Barometer", "Mean Frequency: %.1f Hz", freq);
@@ -8747,7 +8848,7 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
     stat.add("Barometer Sensor", "N/A");
   }
 
-  if (sl_tools::isZED2OrZED2i(mCamRealModel)) {
+  if (!mSvoMode && !mSimMode && sl_tools::isZED2OrZED2i(mCamRealModel)) {
     stat.addf("Left CMOS Temp.", "%.1f °C", mTempLeft);
     stat.addf("Right CMOS Temp.", "%.1f °C", mTempRight);
 
@@ -8759,7 +8860,7 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
     stat.add("Right CMOS Temp.", "N/A");
   }
 
-  if (sl_tools::isZEDX(mCamRealModel)) {
+  if (!mSvoMode && !mSimMode && sl_tools::isZEDX(mCamRealModel)) {
     stat.addf("Camera Temp.", "%.1f °C", mTempImu);
 
     if (mTempImu > 70.f) {
@@ -8983,9 +9084,9 @@ void ZedCamera::callback_clickedPoint(const geometry_msgs::msg::PointStamped::Sh
       camZ);
   } catch (tf2::TransformException & ex) {
     rclcpp::Clock steady_clock(RCL_STEADY_TIME);
-    RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1.0, "Transform error: %s", ex.what());
+    RCLCPP_WARN_THROTTLE(get_logger(), steady_clock, 1000.0, "Transform error: %s", ex.what());
     RCLCPP_WARN_THROTTLE(
-      get_logger(), steady_clock, 1.0,
+      get_logger(), steady_clock, 1000.0,
       "The tf from '%s' to '%s' is not available.",
       msg->header.frame_id.c_str(), mLeftCamOptFrameId.c_str());
 
@@ -9414,6 +9515,31 @@ void ZedCamera::callback_fromLL(
     ll_pt.getLatitude(false), ll_pt.getLongitude(false), ll_pt.getAltitude(),
     res->map_point.x, res->map_point.y, res->map_point.z);
 }
+
+void ZedCamera::callback_clock(const rosgraph_msgs::msg::Clock::SharedPtr msg)
+{
+  DEBUG_SIM("*** CLOCK CALLBACK ***");
+  rclcpp::Time msg_time(msg->clock, RCL_ROS_TIME);
+
+  try {
+    if (msg_time != mLastClock) {
+      mClockAvailable = true;
+      DEBUG_SIM("Received an updated '/clock' message.");
+    } else {
+      mClockAvailable = false;
+      DEBUG_SIM("Received a NOT updated '/clock' message.");
+    }
+    mLastClock = msg_time;
+  } catch (...) {
+    RCLCPP_WARN_STREAM(
+      get_logger(), "Error comparing clock messages: " <<
+        static_cast<int>(msg_time.get_clock_type()) << " vs " <<
+        static_cast<int>(mLastClock.get_clock_type()));
+
+    mClockAvailable = false;
+  }
+}
+
 
 }  // namespace stereolabs